Enum postgres_large_object::Mode

pub enum Mode {
    Read,
    Write,
    ReadWrite,
}

Large object access modes.

Note that Postgres currently does not make any distinction between the Write and ReadWrite modes.

Variants

Read

An object opened in this mode may only be read from.

Write

An object opened in this mode may be written to.

ReadWrite

An object opened in this mode may be read from or written to.