Struct postgres_large_object::LargeObject

pub struct LargeObject<'a> {
    // some fields omitted
}

Represents an open large object.

Methods

impl<'a> LargeObject<'a>

fn truncate(&mut self, len: i64) -> Result<()>

Truncates the object to the specified size.

If len is larger than the size of the object, it will be padded with null bytes to the specified size.

fn finish(self) -> Result<()>

Consumes the LargeObject, cleaning up server side state.

Functionally identical to the Drop implementation on LargeObject except that it returns any errors to the caller.