Trait postgres_large_object::LargeObjectTransactionExt

pub trait LargeObjectTransactionExt {
    fn open_large_object<'a>(&'a self, oid: Oid, mode: Mode) -> Result<LargeObject<'a>>;
}

An extension trait adding functionality to open large objects.

Required Methods

fn open_large_object<'a>(&'a self, oid: Oid, mode: Mode) -> Result<LargeObject<'a>>

Opens the large object with the specified Oid in the specified Mode.